Pan/tilt point at nearest plane!
Added htest.py, a file that should already be in rhazzed's fapping repo, but added it here for completeness. Use the fapping repo to run htest.py, (which, in turn, calls aselSend.py) on the PiAware server. Run azelRecv.py on this (the pan/tilt server) so that it can drive the pan/tilt motors to point to the plane that the PiAware says.  (In my case, I want to point at the plane that's the closest to the receiver QTH.)